Added clustering purity module
Module checks the purity of the clusters after clustering using set
hyperparameters. Purity is the measure of the most occuring genre
within a cluster. A cluster is 100% pure if only one genre exists
within it. Module returns the mean, median, and standard deviation
of the purity for all clustering attempted. Module will cluster based
on 2 hyperparameters, which are cluster size and number of features.